inclination angle is deg
Add unit "deg" for the [orbital] "inclination" field in the table
in Section 3.  Up till now the table listed no unit for this field.
But the text in sec 2.3.2 says "angles in degrees" and all existing
EPN-TAP services currently report this column with units of "deg".
So I think this was just an oversight.